+++ openseries/datefixer.py
@@ -1,12 +1,16 @@
 import datetime as dt
 from dateutil.relativedelta import relativedelta
 from holidays import country_holidays, list_supported_countries
-from numpy import array, busdaycalendar, datetime64, is_busday
-from pandas import Timestamp, to_datetime
+from numpy import array, busdaycalendar, datetime64, is_busday, where
+from pandas import date_range, Timestamp, to_datetime
+from pandas.tseries.offsets import CustomBusinessDay
 
 
 def holiday_calendar(
-    startyear: int, endyear: int, countries: list | str = "SE"
+    startyear: int,
+    endyear: int,
+    countries: list | str = "SE",
+    custom_holidays: list | dict | None = None,
 ) -> busdaycalendar:
     """Function to generate a business calendar
 
@@ -18,6 +22,9 @@ def holiday_calendar(
         Last year in date range generated
     countries: list | str, default: "SE"
         (List of) country code(s) according to ISO 3166-1 alpha-2
+    custom_holidays: list | dict | None, default: None
+        Argument where missing holidays can be added as
+        {"2021-02-12": "Jack's birthday"} or ["2021-02-12"]
 
     Returns
     -------
@@ -31,18 +38,19 @@ def holiday_calendar(
     years = [y for y in range(startyear, endyear)]
 
     if isinstance(countries, str) and countries in list_supported_countries():
-        hols = array(
-            sorted(country_holidays(country=countries, years=years).keys()),
-            dtype="datetime64[D]",
-        )
+        staging = country_holidays(country=countries, years=years)
+        if custom_holidays is not None:
+            staging.update(custom_holidays)
+        hols = array(sorted(staging.keys()), dtype="datetime64[D]")
     elif isinstance(countries, list) and all(
         country in list_supported_countries() for country in countries
     ):
         countryholidays = []
-        for country in countries:
-            countryholidays.extend(
-                country_holidays(country=country, years=years).keys()
-            )
+        for i, country in enumerate(countries):
+            staging = country_holidays(country=country, years=years)
+            if i == 0 and custom_holidays is not None:
+                staging.update(custom_holidays)
+            countryholidays.extend(staging.keys())
         hols = array(sorted(list(set(countryholidays))), dtype="datetime64[D]")
     else:
         raise Exception(
@@ -83,10 +91,11 @@ def date_fix(d: str | dt.date | dt.datetime | datetime64 | Timestamp) -> dt.date
 
 def date_offset_foll(
     raw_date: str | dt.date | dt.datetime | datetime64 | Timestamp,
-    countries: str | list = "SE",
     months_offset: int = 12,
     adjust: bool = False,
     following: bool = True,
+    countries: str | list = "SE",
+    custom_holidays: list | dict | None = None,
 ) -> dt.date:
     """Function to offset dates according to a given calendar
 
@@ -95,14 +104,17 @@ def date_offset_foll(
     raw_date: str | datetime.date | datetime.datetime | numpy.datetime64 |
     pandas.Timestamp
         The date to offset from
-    countries: list | str, default: "SE"
-        (List of) country code(s) according to ISO 3166-1 alpha-2
     months_offset: int, default: 12
         Number of months as integer
     adjust: bool, default: False
         Determines if offset should adjust for business days
     following: bool, default: True
         Determines if days should be offset forward (following) or backward
+    countries: list | str, default: "SE"
+        (List of) country code(s) according to ISO 3166-1 alpha-2
+    custom_holidays: list | dict | None, default: None
+        Argument where missing holidays can be added as
+        {"2023-02-12": "Jack's birthday"} or ["2023-02-12"]
 
     Returns
     -------
@@ -124,7 +136,10 @@ def date_offset_foll(
         startyear = min([raw_date.year, new_date.year])
         endyear = max([raw_date.year, new_date.year])
         calendar = holiday_calendar(
-            startyear=startyear, endyear=endyear, countries=countries
+            startyear=startyear,
+            endyear=endyear,
+            countries=countries,
+            custom_holidays=custom_holidays,
         )
         while not is_busday(dates=new_date, busdaycal=calendar):
             new_date += day_delta
@@ -135,6 +150,7 @@ def date_offset_foll(
 def get_previous_business_day_before_today(
     today: dt.date | None = None,
     countries: str | list = "SE",
+    custom_holidays: list | dict | None = None,
 ):
     """Function to bump backwards to find the previous business day before today
 
@@ -144,10 +160,14 @@ def get_previous_business_day_before_today(
         Manual input of the day from where the previous business day is found
     countries: list | str, default: "SE"
         (List of) country code(s) according to ISO 3166-1 alpha-2
+    custom_holidays: list | dict | None, default: None
+        Argument where missing holidays can be added as
+        {"2023-02-12": "Jack's birthday"} or ["2023-02-12"]
+
     Returns
     -------
     datetime.date
-        The previous Swedish business day
+        The previous business day
     """
 
     if today is None:
@@ -156,7 +176,66 @@ def get_previous_business_day_before_today(
     return date_offset_foll(
         today - dt.timedelta(days=1),
         countries=countries,
+        custom_holidays=custom_holidays,
         months_offset=0,
         adjust=True,
         following=False,
     )
+
+
+def offset_business_days(
+    ddate: dt.date,
+    days: int,
+    countries: list | str = "SE",
+    custom_holidays: list | dict | None = None,
+) -> dt.date:
+    """Function to bump a date by business days instead of calendar days
+
+    Parameters
+    ----------
+    ddate: datetime.date
+        Manual input of the day from where the previous business day is found
+    days: int
+        The number of business days to offset the original date
+    countries: list | str, default: "SE"
+        (List of) country code(s) according to ISO 3166-1 alpha-2
+    custom_holidays: list | dict | None, default: None
+        Argument where missing holidays can be added as
+        {"2023-02-12": "Jack's birthday"} or ["2023-02-12"]
+
+    Returns
+    -------
+    datetime.date
+        The new offset business day
+    """
+    ndate = ddate + dt.timedelta(days=days)
+    if days < 0:
+        calendar = holiday_calendar(
+            startyear=ndate.year,
+            endyear=ddate.year,
+            countries=countries,
+            custom_holidays=custom_holidays,
+        )
+        local_bdays = date_range(
+            periods=abs(days + 10), end=ddate, freq=CustomBusinessDay(calendar=calendar)
+        ).date
+    else:
+        calendar = holiday_calendar(
+            startyear=ddate.year,
+            endyear=ndate.year,
+            countries=countries,
+            custom_holidays=custom_holidays,
+        )
+        local_bdays = date_range(
+            start=ddate, periods=days + 10, freq=CustomBusinessDay(calendar=calendar)
+        ).date
+
+    while ddate not in local_bdays:
+        if days < 0:
+            ddate -= dt.timedelta(days=1)
+        else:
+            ddate += dt.timedelta(days=1)
+
+    idx = where(local_bdays == ddate)[0]
+
+    return local_bdays[idx + days][0]