+++ CHANGELOG.md
@@ -7,6 +7,12 @@ the resources to issue any form
 of guarantee that this log will cover all changes, and I will not attempt to go back
 very far in history.
 
+## Version [0.10.3] - 2023-02-12
+
+Added an offset_business_days function in the datefixer module. It bumps/offsets with
+a given number of business days, not calendar days. Also added the ability to add
+custom holidays in the relevant datefixer functions.
+
 ## Version [0.10.2] - 2023-02-11
 
 Added ability to use multiple countries calendars in combination.